requirements.txt
# Core Framework
django==5.2
psycopg # Recommended over psycopg2-binary for production
djangorestframework
django-environ
celery
redis # For Celery broker/backend & caching
graphene-django # If using GraphQL
elasticsearch-dsl # If using Elasticsearch

# Media Library (FR 4.4)
Pillow # For image processing

# API Enhancements & User Management (FR 4.5)
django-filter # For advanced API filtering
djangorestframework-simplejwt # For JWT authentication
requests # For sending webhooks

# Add others as needed: django-mptt, django-reversion, etc.
